docs: update and centralize profile links
Updated Instagram, LinkedIn, GitHub, and portfolio links. Centralized
all social URLs into `SOCIAL_LINKS` constant to ensure a single source
of truth. Cleaned up redundant data structures in `ProfileData` and
`ArielProfile`.

- Updated Instagram to onasafari
- Updated LinkedIn with skipRedirect and new handle
- Updated GitHub to arii
- Updated portfolio to arii.github.io
- Centralized links in src/config/constants.ts
- Removed redundant links array from ProfileData
- Updated ArielProfile to derive links from connect section
