--- src/config/social.ts
@@ -1,6 +1,0 @@
-export const SOCIAL_LINKS = {
-  INSTAGRAM: 'https://www.instagram.com/onasafari/',
-  LINKEDIN: 'https://www.linkedin.com/in/ariel-anders/?skipRedirect=true',
-  GITHUB: 'https://github.com/arii',
-  PORTFOLIO: 'https://arii.github.io'
-} as const;
